+++ grpc/grpc/workflows/init-server.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest";
-import { InitWorkflowDefinition } from "./init.ts";
+import { InitWorkflowDefinition } from "./init-server.ts";
 import { runWorkflow } from "@saflib/workflows";
 
-describe("init", () => {
+describe("grpc/init-server", () => {
   it("should successfully dry run", async () => {
     const result = await runWorkflow({
       definition: InitWorkflowDefinition,
